requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
trl>=0.7.0
flash-attn>=2.5.0  # NEW: Flash Attention 2 for faster training

# Vector Database & RAG
chromadb>=0.4.18
langchain>=0.1.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rank-bm25>=0.2.2  # NEW: BM25 for sparse retrieval

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-pptx>=0.6.21
python-docx>=1.1.0
PyMuPDF>=1.23.0  # fitz for PDF processing

# Evaluation & Metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.1
evaluate>=0.4.0
bert-score>=0.3.13  # NEW: BERTScore for semantic evaluation

# Reranking & Cross-encoders
cross-encoder>=0.1.0  # NEW: For reranking retrieved documents

# OCR & Vision-Language Models
pytesseract>=0.3.10  # NEW: OCR for extracting text from images
Pillow>=10.0.0  # Image processing
opencv-python>=4.8.0  # NEW: Computer vision for image preprocessing
# NOTE: Tesseract-OCR must be installed separately on the system
# Windows: choco install tesseract
# Linux: sudo apt-get install tesseract-ocr
# macOS: brew install tesseract

# API Integration
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
arxiv>=2.0.0
requests>=2.31.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1
omegaconf>=2.3.0

# Visualization & Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.15.0
wandb>=0.16.0
tqdm>=4.66.0

# Web Interface (Optional)
gradio>=4.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
rich>=13.7.0
click>=8.1.7
jsonlines>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
